# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/incidents/": {
            "get": {
                "tags": [
                    "incidents"
                ],
                "summary": "List Incidents",
                "operationId": "list_incidents_incidents__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListIncidentsResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "incidents"
                ],
                "summary": "Create Incident",
                "description": ":param request: Incoming HTTP request object\n:param create_request: Data required to create a new incident, containing title, description, created_at, and location\n:param user_payload: Payload containing user information, used here to extract the sender's user ID\n:return: Response object containing the ID of the newly created incident",
                "operationId": "create_incident_incidents__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIncidentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateIncidentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/incidents/{incident_id}": {
            "delete": {
                "tags": [
                    "incidents"
                ],
                "summary": "Delete Incident",
                "description": ":param request: The HTTP request object.\n:param incident_id: The ID of the incident to be deleted.\n:param user_payload: The payload containing user information, with a default value of None.\n:return: An instance of DeleteIncidentResponse indicating the result of the delete operation.",
                "operationId": "delete_incident_incidents__incident_id__delete",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Incident Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteIncidentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "incidents"
                ],
                "summary": "Edit Incident Data",
                "description": ":param request: The request object containing metadata about the request.\n:param incident_id: The unique identifier of the incident to be edited.\n:param edit_request: An object containing the new data for the incident including title, description, location, and updated_at values.\n:param user_payload: An optional payload that includes user information such as user_id. Defaults to None.\n:return: An `EditIncidentDataResponse` object indicating the result of the operation.",
                "operationId": "edit_incident_data_incidents__incident_id__put",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Incident Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditIncidentDataRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EditIncidentDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/incidents/{incident_id}/authorize": {
            "post": {
                "tags": [
                    "incidents"
                ],
                "summary": "Authorize Incident",
                "description": ":param request: The request object containing the HTTP request details.\n:param incident_id: An integer representing the unique identifier of the incident to be authorized.\n:param auth_request: An object of type AuthorizeIncidentRequest containing the authorization request data.\n:param user_payload: A dictionary containing user-related data extracted from the JWT token. Default is None.\n:return: An instance of AuthorizeIncidentResponse indicating the result of the authorization process.",
                "operationId": "authorize_incident_incidents__incident_id__authorize_post",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Incident Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthorizeIncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorizeIncidentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/incidents/{incident_id}/vote": {
            "put": {
                "tags": [
                    "incidents"
                ],
                "summary": "Incident Vote",
                "description": "Parameters\n----------\nrequest : Request\n    The HTTP request object.\nincident_id : int\n    The unique identifier of the incident to vote on.\nvote_request : IncidentVoteRequest\n    The request body containing the user's vote.\nuser_payload : dict, optional\n    The payload containing user information extracted from the JWT token (default is None).\n\nReturns\n-------\nAuthorizeIncidentResponse\n    Response indicating the result of the voting action.\n\nRaises\n------\nHTTPException\n    If the incident with the specified ID does not exist.",
                "operationId": "incident_vote_incidents__incident_id__vote_put",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Incident Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IncidentVoteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorizeIncidentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "incidents"
                ],
                "summary": "Get Incident Vote",
                "description": "Parameters\n----------\nrequest : Request\n    The HTTP request object.\nincident_id : int\n    The unique identifier of the incident to vote on.\nuser_payload : dict, optional\n    The payload containing user information extracted from the JWT token (default is None).\n\nReturns\n-------\nAuthorizeIncidentResponse\n    Response indicating the result whether the incident is liked.\n\nRaises\n------\nHTTPException\n    If the incident with the specified ID does not exist.",
                "operationId": "get_incident_vote_incidents__incident_id__vote_get",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "incident_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Incident Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentIsLiked"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthorizeIncidentRequest": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "AuthorizeIncidentRequest"
            },
            "AuthorizeIncidentResponse": {
                "properties": {},
                "type": "object",
                "title": "AuthorizeIncidentResponse"
            },
...
            "CreateIncidentRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Updated At"
                    },
                    "location": {
                        "type": "string",
                        "title": "Location"
                    },
                    "image_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Id"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "description",
                    "created_at",
                    "location"
                ],
                "title": "CreateIncidentRequest",
                "description": "Represents a request to create a new incident.\n\nAttributes:\n    title (str): The title of the incident.\n    description (str): Detailed description of the incident.\n    created_at (datetime): The timestamp when the incident was created.\n    updated_at (datetime | None): The timestamp when the incident was last updated, can be None.\n    location (str): The location where the incident occurred.\n    image_id (int | None): The ID of the associated image, can be None."
            },
            "CreateIncidentResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "id"
                ],
                "title": "CreateIncidentResponse"
            },
...
            "DeleteIncidentResponse": {
                "properties": {},
                "type": "object",
                "title": "DeleteIncidentResponse"
            },
...
            "EditIncidentDataResponse": {
                "properties": {},
                "type": "object",
                "title": "EditIncidentDataResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "IncidentVoteRequest": {
                "properties": {
                    "vote": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/IncidentVote"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "IncidentVoteRequest"
            },
...
            "ListIncidentsResponse": {
                "properties": {
                    "incidents": {
                        "items": {
                            "$ref": "#/components/schemas/Incident"
                        },
                        "type": "array",
                        "title": "Incidents"
                    }
                },
                "type": "object",
                "required": [
                    "incidents"
                ],
                "title": "ListIncidentsResponse"
            },
...
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "description": "Bearer token",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}